@@ -5,16 +5,16 @@ build-backend = "pdm.backend"
 [project]
 authors = []
 license = { text = "MIT" }
-requires-python = ">=3.9"
+requires-python = ">=3.10.0,<4.0.0"
 dependencies = [
-    "langchain-core<1.0.0,>=0.3.72",
-    "langchain-text-splitters<1.0.0,>=0.3.9",
-    "langsmith>=0.1.17",
-    "pydantic<3.0.0,>=2.7.4",
-    "SQLAlchemy<3,>=1.4",
-    "requests<3,>=2",
-    "PyYAML>=5.3",
-    "async-timeout<5.0.0,>=4.0.0; python_version < \"3.11\"",
+    "langchain-core>=0.3.72,<1.0.0",
+    "langchain-text-splitters>=0.3.9,<1.0.0",
+    "langsmith>=0.1.17,<1.0.0",
+    "pydantic>=2.7.4,<3.0.0",
+    "SQLAlchemy>=1.4.0,<3.0.0",
+    "requests>=2.0.0,<3.0.0",
+    "PyYAML>=5.3.0,<7.0.0",
+    "async-timeout>=4.0.0,<5.0.0; python_version < \"3.11\"",
 ]
 name = "langchain"
 version = "0.3.27"
